plugins: abrt-journal-core: Don’t assume anything about uid_t
Sure, maybe glibc defines uid_t as a uint32_t, but POSIX does not impose
any limits on the underlying type, so let’s just accept whatever systemd
gives us. If what systemd gives us is non-compliant, then we are already
in trouble and it makes no sense to defend against nasal demons.
Moreover, using a long for bounds and passing UINT_MAX is just asking
for trouble (i.e. it fails on 32-bit systems).

+++ src/plugins/abrt-journal.h
@@ -54,7 +54,7 @@ int abrt_journal_get_field(abrt_journal_t *journal,
 
 int abrt_journal_get_int_field(abrt_journal_t *journal,
                                const char *field,
-                               int *value);
+                               long *value);
 
 int abrt_journal_get_unsigned_field(abrt_journal_t *journal,
                                     const char *field,
